+++ src/options/types.ts
@@ -1,11 +1,3 @@
-export enum OptionType {
-    PRE_METHOD_ADVICE,
-    PRE_MIXING_ADVICE,
-    POST_METHOD_ADVICE,
-    POST_MIXING_ADVICE,
-    OMITS
-}
-
 interface OptionKey {
     // the primary internal identifier for an option key;
     // must be kept in sync with the type definition MixOptions
@@ -17,7 +9,7 @@ const PRE_METHOD_ADVICE_KEY: OptionKey = {
     primary: 'pre_method_advice',
     aliases: [
         'before_hook',
-        'hook_before'
+        'hook_before',
     ],
 };
 
@@ -25,7 +17,7 @@ const PRE_MIXING_ADVICE_KEY: OptionKey = {
     primary: 'pre_mixing_advice',
     aliases: [
         'premixing_hook',
-        'premixing', 
+        'premixing',
         'premix'
     ],
 };
@@ -42,9 +34,9 @@ const POST_MIXING_ADVICE_KEY: OptionKey = {
     primary: 'post_mixing_advice',
     aliases: [
         'postmixing_hook',
-        'postmixing', 
+        'postmixing',
         'postmix'
-    ], 
+    ],
 };
 
 const OMITS_KEY: OptionKey = {
@@ -52,11 +44,18 @@ const OMITS_KEY: OptionKey = {
     aliases: []
 };
 
-let specifies_option = (option_key: OptionKey, key: string) => {
-    return (option_key.primary === key) || (key in option_key.aliases);
-};
+export enum OptionType {
+    PRE_METHOD_ADVICE,
+    PRE_MIXING_ADVICE,
+    POST_METHOD_ADVICE,
+    POST_MIXING_ADVICE,
+    OMITS
+}
 
 export var option_type_of = (key: string): OptionType => {
+    let specifies_option = (option_key: OptionKey, key: string) => {
+        return (option_key.primary === key) || (key in option_key.aliases);
+    };
     if (specifies_option(PRE_METHOD_ADVICE_KEY, key)) {
         return OptionType.PRE_MIXING_ADVICE;
     } else if (specifies_option(PRE_MIXING_ADVICE_KEY, key)) {