+++ src/index.ts
@@ -1,0 +1,5 @@
+import {make_mixin} from './mixin';
+import {mix} from './mixer';
+
+
+export {make_mixin, mix};